// packages/dnb-eufemia/src/extensions/forms/Field/String/String.d.ts
import React from 'react';
import type { InputProps } from '../../../../components/input/Input';
import type { InputMaskedProps } from '../../../../components/InputMasked';
import type { TextareaProps } from '../../../../components/Textarea';
import type { FieldBlockWidth } from '../../FieldBlock';
import type { TextCounterProps } from '../../../../fragments/TextCounter';
import type { FieldProps } from '../../types';
export type Props = FieldProps<string, undefined | string> & {
    multiline?: boolean;
    inputClassName?: string;
    ref?: React.RefObject<HTMLInputElement | HTMLTextAreaElement>;
    width?: FieldBlockWidth;
    size?: InputProps['size'] | TextareaProps['size'];
    keepPlaceholder?: InputProps['keepPlaceholder'];
    minLength?: number;
    maxLength?: number;
    pattern?: string;
    type?: InputProps['type'];
    align?: InputProps['align'];
    selectAll?: InputProps['selectAll'];
    clear?: boolean;
    mask?: InputMaskedProps['mask'];
    leftIcon?: string;
    rightIcon?: string;
    submitElement?: InputProps['submitElement'];
    capitalize?: boolean;
    trim?: boolean;
    rows?: TextareaProps['rows'];
    autoResizeMaxRows?: TextareaProps['autoResizeMaxRows'];
    autoResize?: TextareaProps['autoResize'];
    characterCounter?: Omit<TextCounterProps, 'text'> | number;
    autoComplete?: HTMLInputElement['autocomplete'];
    inputMode?: React.HTMLAttributes<HTMLInputElement>['inputMode'];
    autoCorrect?: React.HTMLAttributes<HTMLInputElement>['autoCorrect'];
    spellCheck?: React.HTMLAttributes<HTMLInputElement>['spellCheck'];
    autoFocus?: React.HTMLAttributes<HTMLInputElement>['autoFocus'];
    autoCapitalize?: React.HTMLAttributes<HTMLInputElement>['autoCapitalize'];
    onKeyDown?: React.KeyboardEventHandler<HTMLInputElement>;
};
declare function StringComponent(props: Props): import("react/jsx-runtime").JSX.Element;
export default StringComponent;